separate out wrp parsing functions from boot time parser
